create separate price interface
