+++ server/src/dto/Price.ts
@@ -1,0 +1,11 @@
+import { IPrice } from "@interfaces/Price";
+
+export class PriceDTOConstructor implements IPrice {
+    current: number;
+    old: number;
+
+    constructor({ current, old }: IPrice) {
+        this.current = current;
+        this.old = old < current ? null : old;
+    }
+}
