+++ server/src/interfaces/Price.ts
@@ -1,0 +1,4 @@
+export interface IPrice {
+    current: number
+    old: number | null
+}
